#  Apply bin optimizations to msgpack_unpacker