--- a/rebar.config
+++ b/rebar.config
@@ -18,7 +18,7 @@
 
 {profiles,
  [{test,
-   [{plugins, [rebar3_eqc]}, {deps, [{"jiffy", "1.1.1"}]}]
+   [{plugins, [rebar3_eqc]}]
   }]}.
 
 %% {port_sources, ["c_src/*.c"]}.