The optimizations give a 15% speed bump

**Before the change:**
```
Erlang/OTP 24 [erts-12.3.2.2] [source] [64-bit] [smp:10:10] [ds:10:10:10] [async-threads:1]

Eshell V12.3.2.2  (abort with ^G)
1> msgpack_bench:bench().
Done 1000 iterations, took 66 ms
2> msgpack_bench:bench(100000).
Done 100000 iterations, took 15669 ms
3> msgpack_bench:bench(10000000).
Done 10000000 iterations, took 865086 ms
```

**After the change:**
```
 Erlang/OTP 24 [erts-12.3.2.2] [source] [64-bit] [smp:10:10] [ds:10:10:10] [async-threads:1]

Eshell V12.3.2.2  (abort with ^G)
1> msgpack_bench:bench().
Done 1000 iterations, took 53 ms
2> msgpack_bench:bench(100000).
Done 100000 iterations, took 13041 ms
3> msgpack_bench:bench(10000000).
Done 10000000 iterations, took 747537 ms
4>

```

**Benchmark code:**

```erl
-module(msgpack_bench).

-export([bench/0, bench/1]).

bench() ->
    bench(1000).

bench(Iterations) ->
    {ok, TestData} = file:read_file("/tmp/pokedex.json"),
    TestDataDecoded = jiffy:decode(TestData),
    TestDataPacked = msgpack:pack(TestDataDecoded),
    {Time, _} = timer:tc(fun() -> [ begin
        msgpack:unpack(TestDataPacked),
        ok
    end || _ <- lists:seq(0, Iterations)]
    end),

    io:format("Done ~p iterations, took ~p ms", [Iterations, Time]).

```

Json used for the benchmark: https://raw.githubusercontent.com/Biuni/PokemonGO-Pokedex/master/pokedex.json