@raz-adroll 2022-08-10
Shared by @tak30 on how these optimizations work:
https://www.erlang.org/doc/efficiency_guide/binaryhandling.html
https://rhye.org/post/erlang-binary-matching-performance/