@tak30 2022-08-11
> Can you put comments on those binary matches (at least in some of them) so future me doesn't say _"hey… we can simplify this by removing the `/binary` pieces since it's the same anyway"_? 🙏🏻

Done!